move fn to utils
