+++ src/utils/misc.ts
@@ -100,3 +100,16 @@ export function getLokiUrl(
 
   return loki_url;
 }
+
+
+export function getRandom(arr: string[], n: number) {
+  let result = new Array(n),
+    len = arr.length,
+    taken = new Array(len);
+  while (n--) {
+    let x = Math.floor(Math.random() * len);
+    result[n] = arr[x in taken ? taken[x] : x];
+    taken[x] = --len in taken ? taken[len] : len;
+  }
+  return result;
+}
